New endpoint to receive device monitoring metrics
New endpoint `POST :server/api/v1/devices/:uuid/monitoring` accepting
a json body.

The body is parsed and known metrics are extracted. The extracted
metrics are saved to one timescaledb hypertable and the raw json is
saved to a different table.

All metrics are optional, but if the server cannot at least parse the
date then the request is rejected and a 400 Bad Request is returned.