New endpoint PATCH devices/:device-uuid to partially update a device
Receives requests with the following body:

```
{
    "deviceName": "some device name",
    "notes": "Device notes"
}

```

If any of the attributes is missing, it will not be updated.

The previously existing endpoint `PUT devices/:device-uuid`, now also
receives this json payload, however, if `notes` is missing, that
attribute is deleted. `deviceName` is a required attribute and
therefore the PUT request will fail if it's not set.