Upgrade libats, slick
Bumping libats requires bumping slick. The new version of slick comes
with java.time.Instant column mappers, which were not available in
previous versions.

These mappers use a more precise representation to store Instant in
MySQL. They rely on converting an Instant to a string and then using a
VARCHAR instead of a DATETIME and then parsing the string into a date
when reading.

For legacy reasons, we use mappers (javaInstantMapping) that convert
an Instant to a java.sql.Timestamp than is then handled by jdbc and
saved to a DATETIME column.

Since our date columns are all using DATETIME(3) instead of VARCHAR,
when inserting values the operation fails, since jdbc is trying to
insert a string into a DATETIME.

To fix this and use the default slick mappers, we would need first to
change the columns from DATETIME to VARCHAR, without losing
data. However, it is not clear if this is the correct solution for our
services, and this would take significant time to test and run
correctly.

Therefore, this commit explicitly instructs slick to use
javaInstantMapping, instead of its default mappings, when defining the
tables for slick.

However, when writing queries in the slick DSL/methods (.filter,
<,>,===, etc), it's not possible to explicitly make slick use the non
default mappers (javaInstantMapping), therefore we convert some
queries to raw sql where we can use prepared statements and explicitly
set the values for Instant values.