Add api to set device hibernation status
In the case of a group, this implementation is not correct and should
not be merged like this.

If the database operation fails inbetween batches, some devices will
be moved to hibernation and some will not.

In addition, if some operation fails, some messages might not be
published to the message bus, even if other operations were already
published.